requirements.txt
# Core ML Libraries (update torch based on your setup below)
# See INSTALLATION.md for CUDA-specific installation instructions

# For CPU-only (development/testing)
# torch==2.3.1
# torchvision==0.18.1

# For CUDA 11.8 (cu118)
# torch==2.3.1+cu118
# torchvision==0.18.1+cu118

# For CUDA 12.4/12.5 (cu124/cu125)
# torch==2.4.1+cu124
# torchvision==0.19.1+cu124

# Audio and Vision Libraries
transformers==4.35.2
librosa==0.10.0
numpy==1.24.3
Pillow==10.0.0
soundfile==0.12.1

# Vision-Language Model (for CLIP)
# Install via: pip install git+https://github.com/openai/CLIP.git
# Or: pip install clip-by-openai==1.3.0
